World descriptors are now filled when a new world is created.

+++ src/main/java/hu/unideb/inf/prt/petriDish/WorldDescriptor.java
@@ -27,7 +27,7 @@ public class WorldDescriptor {
 	@XmlElement(name = "genotype")
 	private List<Genotype> genotypes;
 	@XmlElement
-	private int fitness;
+	private long fitness;
 	@XmlElement
 	private GameConfiguration conf;
 
@@ -72,11 +72,11 @@ public class WorldDescriptor {
 		}
 	}
 
-	public int getFitness() {
+	public long getFitness() {
 		return fitness;
 	}
 
-	public void setFitness(int f) {
+	public void setFitness(long f) {
 		fitness = f;
 	}
